[TEST] Fixed failed tests and split UT/IT
