+++ test-src/org/pentaho/common/ui/TestUserSettingService.java
@@ -15,7 +15,7 @@
  * Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
  */
 
-package org.pentaho.common.ui.test;
+package org.pentaho.common.ui;
 
 import java.util.HashMap;
 import java.util.List;
