+++ test-src/org/pentaho/common/ui/messages/MessagesIT.java
@@ -15,7 +15,7 @@
  * Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
  */
 
-package org.pentaho.common.ui.test;
+package org.pentaho.common.ui.messages;
 
 import static org.junit.Assert.assertEquals;
 
@@ -28,13 +28,11 @@ import org.pentaho.platform.plugin.services.pluginmgr.PluginClassLoader;
 import org.pentaho.platform.plugin.services.pluginmgr.PluginResourceLoader;
 import org.pentaho.test.platform.engine.core.MicroPlatform;
 
-import org.pentaho.common.ui.messages.Messages;
-
 @SuppressWarnings( "nls" )
-public class MessagesTest {
-  MicroPlatform mp;
+public class MessagesIT {
+  private MicroPlatform mp;
 
-  public MessagesTest() throws ClassNotFoundException, ObjectFactoryException {
+  public MessagesIT() throws ClassNotFoundException, ObjectFactoryException {
     mp = new MicroPlatform( "" );
   }
 
@@ -48,10 +46,7 @@ public class MessagesTest {
 
   @Test
   public void testMessages() {
-
     mp.define( IPluginResourceLoader.class, TstPluginResourceLoader.class );
-
-    new Messages();
 
     assertEquals( "Wrong message returned", "test message", Messages.getString( "TEST.MESSAGE1" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
 
@@ -64,12 +59,10 @@ public class MessagesTest {
 
     assertEquals(
         "Wrong message returned", "test message 5: A B C D", Messages.getString( "TEST.MESSAGE5", "A", "B", "C", "D" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$ //$NON-NLS-6$ //$NON-NLS-7$
-
   }
 
   @Test
   public void testErrorMessages() {
-
     assertEquals(
         "Wrong message returned", "TEST.ERROR_0001 - test error 1", Messages.getErrorString( "TEST.ERROR_0001" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
 
@@ -81,22 +74,18 @@ public class MessagesTest {
 
     assertEquals(
         "Wrong message returned", "TEST.ERROR_0004 - test error 4: A B C", Messages.getErrorString( "TEST.ERROR_0004", "A", "B", "C" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$ //$NON-NLS-6$
-
   }
 
   @Test
   public void testBadKey() {
-
     assertEquals( "Wrong message returned", "!bogus key!", Messages.getString( "bogus key" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
 
     assertEquals(
         "Wrong message returned", "test.ERROR_0001 - !test.ERROR_0001_BOGUS!", Messages.getErrorString( "test.ERROR_0001_BOGUS" ) ); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-
   }
 
   @Test
   public void testBadEncoding() {
-
     // the messages.properties file has a bad encoding for the test.encode1 property, this causes a
     // MissingResourceException which
     // manifests as a returned string of !<key>! for all getString calls including the good strings
