+++ test-src/org/pentaho/common/ui/metadata/service/TestModelProvider.java
@@ -15,7 +15,7 @@
  * Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
  */
 
-package org.pentaho.common.ui.test;
+package org.pentaho.common.ui.metadata.service;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
