+++ test-src/org/pentaho/common/ui/models/TreeBrowserMapperTest.java
@@ -15,7 +15,7 @@
  * Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
  */
 
-package org.pentaho.common.ui.test;
+package org.pentaho.common.ui.models;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Before;
