+++ test-src/org/pentaho/common/ui/models/TreeBrowserModelUtilsTest.java
@@ -1,4 +1,4 @@
-package org.pentaho.common.ui.test;
+package org.pentaho.common.ui.models;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Before;
